requirements.txt
Django==2.2.9
djangorestframework==3.11.0
requests==2.22.0

